e2e: replace WaitEtcdTPRReady with WaitUntilOperatorReady
